Language improvement in challenge
The first challenge in the novice Python 01 lesson asks why the
vertical lines in the plot are not vertical. This is a little
confusing since we are describing the lines as vertical and then
saying that they aren't. This changes the language to ask why
the lines are not "perfectly vertical" to avoid any confusion.