Remove reference to strides in natural numbers exercise
The ability to skip over count in numbers other than one isn't necessary
for the exercise, so this presents the minimal amount of information
that the student needs.

This also allows us to remove discussion of strides from the discussion
of indexing in the 01 notebook letting us minimize the amount of extra
material we are delivering.