Add transition text
The part that describes assignment jumps from talking about arrays to scalars and back to arrays without any transitioning text.  Transitioning text is added at the beginning of the discussion about assignment, and revised transitioning text is added at the end of that sub-topic, to make the switch less jarring.