Change order of output and error and corrects error message in 02-loop.md
This commit changes the order of the output after executing the print
statements to demonstrate an "out-of-index" error in 02-loop.md.

The error message that followed the print statement also refers to
a function `print_elements` that is not defined (at the moment) in this lesson.

As this lesson (02-loop.md) now comes before "Creating functions" (06-func.md)
I have assumed that we are interested in removing references to
`print_elements`.