Removal for function (too early in lesson program)
In this lesson there is an example that uses a function "def sign(num)" to illustrate some decision making capability. I don't know if this material has been reordered recently, but given that function definitions do not appear until the next lesson, I would argue that this be removed and just shown as a simple block of code that prints whether a number is -ve or +ve. I have suggested a potential replacement.