Update 01-numpy.md
In the challenge 'Sorting out references' multiple assignment is used without any previous mention. I have added the use of assigning multiple values to multiple variables in the numpy methods section. I understand that this is introducing two new concepts at the same time (methods and multiple assignment).

It does allow the student to chunk the three descriptive statistics though so is might lower the load on working memory
The other options are to introduce simultaneous assignment in it's own section, either in this lesson or move it elsewhere and move the challenge. There is also the option of removing the multiple assignment from the challenge and assigning each variable separately.