Update 01-numpy.md
Students in the last session in Oslo were asking how to see what had been declared/loaded already. It is a good command that I often use when working, I think it is easy, and also very nice to know from the start. Showing them the command so early on could help them conceptualise the namespace more effectively.