Add debugging challenge
This exercise has the aim of ensuring learners are able to step through
unseen code with unexpected output to locate issues.  The issues present
are that the loop is not being utilised correctly (height/weight always
set as the first patient's data) and that the height/weight variables
are reversed.