Prefer functions over methods.
Most methods on NumPy arrays also have an equivalent NumPy functions
that takes the array as input. I have noticed that many novice
programmers prefer function (which have a pretty clear input-output
relationship) to methods (which have an implicit input).
This commit changes all instances of calling methods on arrays to
calling the equivalent NumPy function instead. I tried to simplify
the explanation in `01-numpy.md` as well, which hopefully does not
significantly affect the flow of the prose.

I also added a "function" entry to the glossary and linked to it in
`01-numpy.md`.