Update 02-loop.md
I removed text about strings being lists because this isn't accurate.  I've replaced it with a statement that you can access the elements of a string through index or a loop in the same manner.  Avoiding discussion of the difference between a string and a list of characters for now.