Suggest changing list example to string example in  02-loops.md, section "What's in a name"
The original 02-loops section "What's in a name" uses an example using a for loop over a list.
As lists aren't introduced until section 03-lists, I suggest changing this example to a loop over
a string.  Note that there is an excercise "Computing the Value of a Polynomial" in 02-loops that also
uses a list, that may be moved to 03-lists as well.