Swap the order of episodes on lists and loops
It has been decided (swcarpentry/python-novice-inflammation#453)
that it is better to describe Python lists before loops. This is
the first commit that simply renames the episodes to achieve the
desired order.

Further work is necessary to ensure that concepts from the
episode on loops are not used in the episode on lists.