Correct pepper-lists analogy
Fix #458.

For the pepper analogy to hold, where x is a list of lists
represented by a pepper pot full of pepper sachets,
x[0] is a sachet of pepper outside of the pot,
x[0][0] represents the first grain of pepper.

The analogy doesn't quite hold because the pepper grains
aren't ordered, but at least it correctly shows one grain
rather than all of the grains in a heap, which is clearly
incorrect.