use floats instead of integers when introducing functions
Apart from changing the example, I have also changed examples further
below in the episode to floats instead of integers.

Motivation:

1) The previous code produced **wrong results** in Python 2.
In Python 2 these two functions do not produce the same result:

```python
def fahr_to_celsius(temp):
    return ((temp - 32) * (5/9))

def fahr_to_celsius_better(temp):
    return ((temp - 32.0) * (5.0/9.0))
```

I know we use Python 3 but I think it is not a good idea to introduce
code which produces a different result with Python 2 which still many
people and codes use.

2) Using decimals annotates to the reader that the function is expected
to work with floats, not only with integers.  If I see an example using
integers, I assume that the function expects integers. Of course this
may be in the category "personal taste".

3) The example is more consistent with the rest of the episode where in
"Composing Functions" we suddenly switch to floats.