Rename variables in `range_overlap` function
Change `low` and `high` variables to `left` and `right` respectively
since that's what is normally used to call the interval endpoints.

Also I think `lowest` and `highest` are quite confusing since they
are actually calculated as `max` and `min` respectively. I have
therefore changed these to `max_left` and `min_right`.

I've also added an additional `assert` to justify the:
```
if not ranges:
    # ranges is None or an empty list
    return None
```
check in the final exercise solution.

Finally, I've removed an unused `import numpy` in the final exercise.