Fleshed out the schedule a little more to avoid confusion.
There was a bit of confusion about the lengths of the coffee breaks and the goind straight from Coffee to Lunch. I've just added in the sections between to increase clarity and give a default schedule for coffee and lunch breaks. 4 x 80 minute segments broken up by coffee and lunch breaks. Obviously just a guideline...