Remove the word "simple" from 06-func.md
I feel that using the word "simple" can be a bit dismissive. I don't think we can assume that novices would agree that the tests outlined would be simple for them to come up with when they're first starting out with Python. I think this can be reworded to convey a similar meaning without being as dismissive.