Update 06-func.md
Following on from the PR #691, the exercises for this episode need around 15 to 20 mins as well I think. The first 3 are simpler than those that follow, so around 2 mins vs 3 mins or so for the two groups. What's your experience been @maxim-belkin and @annefou ? I'm continuing to prepare for my first Python swc teaching session after checkouts here in Adelaide (AU) this week.