Update 04-lists.md (#763)
* Update 04-lists.md

I think it's good to repeat the explanation of behaviour caused by (attempted) copying and then modifying the list in place. From teaching this several times, I think it would help to notify the learner that the use of append here is a second, similar example to what we saw earlier with the salsa list.

* Update 04-lists.md

Replaced brackets with back-ticks and added comma.