02-numpy: diff(): rename variable, streamline text
Rename variable `a` -> `row_start` and add text relating that to the
`data` array encountered earlier.

The text part about "patient data is _longitudinal_" is redundant since
that answers the question posed later regarding the usage of the `axis`
keyword. It intoduces the usage of `axis`, while the following intro to
diff() first uses a 1d array.