Update 03-matplotlib.md
In correspondence to issue #664

Add a short explanation of the differences between libraries and modules

I agree with the point of introducing the concepts of 'libraries' and 'modules'. Because if not explaining them, there might be confusions to audiences. But I feel that the main thread of this lesson is to talk about Python fundamental concepts and dealing with datasets, and the terminologies of 'libraries' and 'modules' are not that much important. So, as a tradeoff, I choose to add a quick note below the first appearance of the concept 'module', so the audience won't get confused here while still following the main thread of this lesson.