Update 12-cmdline.md
This is for the SWC checkout process. I added text at L24-28 clarifying the goals in the lesson, L85-88 adding detail on the sys library, and L127-128 adding detail on argv. I also suggest explaining the use of 'cat' at L165 to view files vs. editing them. These suggestions are based on points of confusion when I was going through the lesson for the live coding demo.