07-cond.md: Close Enough correction (#839)
`print(abs(a - b) < 0.1 * abs(b))` (and preceding Solution 2) leaves out the edge case where `a` is exactly 10% of `b`. This issue comes down to the interpretation of "within" 10%. Does a number within 10% of another number include the 10% value? 

Assume `b = 100` and `0.1 * b = 10`, then the value 10 is within 10% by definition since 10 is 10% of 100. Now set `a = 100 + 10 = 110` which is `b` plus 10% of `b`. Then `a - b = b + 0.1 * b - b = 0.1 * b`, which is within 10% of b.

The correct test is `abs(a - b) <= 0.1 * abs(b)` to account for this edge case.