Update 01-intro.md
Add clearer definition of what a variable is, this could be done somewhere around line 19. For example: Variables are used to store information to be referenced and manipulated in a computer program. It is helpful to think of variables as containers that hold information. Their sole purpose is to label and store data in memory. This data can then be used throughout your program.