Update 01-intro.md
The assignment of weigtht_kg = 60 and then weight_kg=60.0 could be confusing as the second usage assigning a floating point said to "create a variable" where we are over-writing it's value.  The proposed change is intended to clarify that without introducing the potentially confusing topic of auto-typing.  It also made a small adjustment to mention that Python can auto-identify types of information.