Revision to Code Sample
I tried to do a live code demo of this exercise during my Carpentries Instructor Training and noticed that this first piece of code will not run as it is currently written, as the variable "data" is not initialised before it is used. I have added the two additional lines of code necessary to make this code sample work. 
(This is part of my Instructor Training checkout process)