01-intro.md: Improve the explanatory text in the "Variables as Sticky Notes" callout of Episode 1 (#928)
As discussed in swcarpentry/python-novice-inflammation#847, the example
in this callout makes it sound like the reason that variables don't
"remember" where their values come from is a consequence of the fact
that variables behave like sticky notes. However, this is inaccurate and
could mislead learners. For the moment, some modification of the
explanatory text has been suggested as a stop-gap until a better example
can be crafted.