Base "middle" on 0-index consistently
This section of the lesson is inconsistent. It uses data[30,20] as a midpoint but states that this is the 30th row and 20th column.

Either the indices should be changed (per PR) and the text updated a little or the text should be changed to reflect that this is the 31st row and 21st column (not 30th and 20th).