Add explicit parameters for numpy.delete
The alternative solution of using numpy.delete() to remove a column from an array is helpful, but the meaning of the arguments is not obvious without explicitly including the names. Even then, "obj" is non-obvious, so I also added a note in the solution description reminding them to use help files if necessary.