Do not mention "exceptions"
The section currently is called 'Errors and Exceptions', but
the section itself only mentions 'errors' and just once
'errors and exceptions'. There might be an argument here about
what the difference is between both, but neither is it currently
made in the material, nor do I think it should in an introductory
course like this.

Because of this, I propose to remove the mention of 'exceptions'
entirely from this course (only 2 places), and especially from a
section heading, to avoid unnecessary confusion by new learners.