remove unnecessary type conversion
This removes the ```float()``` around a (potential) integer division.
Since Python3 the result of an integer division is a floating
point number. This also removes the time and attention capacity
necessary by learners for an explaination of what this does and
why it is there. I do not mean to say that type conversions cannot
be important, but the focus of this excercise is elsewhere.

Note: it was good to have this in place for Python2, but Python2
is dead today and has been for a while. We also require an installation
of Python3 in the setup.