adding labels to plot
While I understand the beauty of having just 3 lines of code to ingest a dataset and output a heatmap, the resulting plot does not meet the basic requirements of an adequate plot. We shouldn't teach people to generate something 'data-sciency' that people can re-use without understanding what is to be seen on the plot.